Issue #377 - use the VMUSB drivers in place rather than as copies
Import madcscaler directly from vmsub/daqconfig rather than a hard to
maintain local copy.
