+++ main/mvlc/Makefile.am
@@ -64,12 +64,13 @@ IMPORTED_MODULES=$(IMPORT_DIR)/CMQDC32RdoHdwr.h \
 	$(IMPORT_DIR)/CDelay.h $(IMPORT_DIR)/CDelay.cpp \
 	$(IMPORT_DIR)/CMesytecBase.cpp $(IMPORT_DIR)/CMesytecBase.h \
 	$(IMPORT_DIR)/CMADC32.cpp $(IMPORT_DIR)/CMADC32.h   \
-	$(IMPORT_DIR)/CMADCChain.h $(IMPORT_DIR)/CMADCChain.cpp 
+	$(IMPORT_DIR)/CMADCChain.h $(IMPORT_DIR)/CMADCChain.cpp  \
+	$(IMPORT_DIR)/CMADCScaler.h $(IMPORT_DIR)/CMADCScaler.cpp
 
 
 LOCAL_MODULES= CMarker.h CMarker.cpp \
 	MADC32Registers.h \
-	CMADCScaler.h CMADCScaler.cpp CMTDC32.h CMTDC32.cpp \
+	CMTDC32.h CMTDC32.cpp \
 	CMDPP16QDC.cpp  CMDPP16QDC.h  CMDPP32QDC.cpp  CMDPP32QDC.h \
 	CMDPP32SCP.cpp  CMDPP32SCP.h  CMDPP.h  CMDPPQDC.h \
 	CNADC2530.h CNADC2530.cpp CV977.h CV977.cpp \
